Refactor code -Delete functions -Rename functions -Move functions and code.
