+++ src/components/index.js
@@ -23,7 +23,9 @@ export {OptionListButton} from './helpers/CardsButton/OptionListButton'
 export {CardTracks} from './TracksList/CardTracks'
 export {setLikeSong} from './helpers/CardsButton/LikeButton'
 export {setFavoritePlaylist} from './helpers/CardsButton/LikeButton'
-export {addLikeTracks} from './Firebase/StoreLikes'
-export {addLikesPlaylists} from './Firebase/StoreLikes'
 export {LoadLikes} from './Firebase/LoadLikes'
-
+export {getImage} from './TracksList/getImage'
+export {HandleChange} from './TrackPlayer/HandleChange'
+export {FormatTime} from './TrackPlayer/FormaTime'
+export {SearchIcons} from './buscador/SearchIcons'
+export {SearchResults} from './buscador/SearchResults'
