+++ src/redux/SpotifyReducer.js
@@ -1,4 +1,3 @@
-
 import {
   SET_TOKEN,
   SET_INFO,
